organize components into folders
- update all imports
