+++ src/pages/_layout.svelte
@@ -1,6 +1,6 @@
 <!-- routify:options preload="proximity" -->
 <script>
-  import Navbar from "../components/Navbar.svelte";
+  import Navbar from "../components/layout/Navbar.svelte";
 </script>
 
 <Navbar />
